combine landing page components into one file
